@@ -8,15 +8,15 @@
     <name>Solace Event Management Agent - Local Storage Plugin</name>
     <description>Solace Event Management Agent - Local Storage Plugin</description>
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <org.json.version>20230227</org.json.version>
-        <camel.version>3.16.0</camel.version>
+        <camel.version>4.0.0-RC2</camel.version>
         <snakeyaml.version>2.0</snakeyaml.version>
         <jacksondatabind.version>2.13.4.2</jacksondatabind.version>
         <junit.version>4.13.2</junit.version>
-        <spring-boot-starter.version>2.7.12</spring-boot-starter.version>
-        <spring-boot-starter-test.version>2.7.12</spring-boot-starter-test.version>
+        <spring-boot-starter.version>3.1.2</spring-boot-starter.version>
+        <spring-boot-starter-test.version>3.1.2</spring-boot-starter-test.version>
         <commons-collections4.version>4.4</commons-collections4.version>
         <jupiter.version>5.3.1</jupiter.version>
         <maven-shade-plugin.version>3.3.0</maven-shade-plugin.version>
