@@ -6,6 +6,7 @@ import com.solace.maas.ep.event.management.agent.plugin.processor.async.AsyncMan
 import com.solace.maas.ep.event.management.agent.plugin.route.manager.AsyncManager;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.camel.CamelContext;
 import org.apache.camel.Exchange;
 import org.apache.camel.ExtendedCamelContext;
 import org.apache.camel.builder.ExchangeBuilder;
@@ -40,7 +41,7 @@ public class AsyncManagerProcessorTests {
                 RouteConstants.SCAN_TYPE, "testScan"
         );
 
-        Exchange exchange = ExchangeBuilder.anExchange(mock(ExtendedCamelContext.class))
+        Exchange exchange = ExchangeBuilder.anExchange((CamelContext) mock(ExtendedCamelContext.class))
                 .withBody(body)
                 .build();
 
